feat: Tenders added, Need to fix its routing from home screen
