+++ server/actions/tenders.ts
@@ -5,26 +5,36 @@
  *
  * Next.js Server Actions for tender form submissions and management.
  * These actions require admin authentication (ccn@nitkkr.ac.in).
+ *
+ * Status is COMPUTED dynamically based on dates:
+ * - 'live': GREATEST(endDate, extendedDate) >= TODAY (today or future)
+ * - 'archived': GREATEST(endDate, extendedDate) < TODAY (past)
  */
 
 import { revalidatePath } from 'next/cache';
+import { eq, sql } from 'drizzle-orm';
 import { z } from 'zod';
 
 import { env } from '~/lib/env/server';
 import { canManageNotifications, getServerAuthSession } from '~/server/auth';
-import {
-  createTender,
-  updateTender,
-  updateTenderExtendedDate,
-  deleteTender,
-  getTenderById,
-  getLiveTenders,
-  getArchivedTenders,
-  getPaginatedTenders,
-  getTenderCount,
-  type TenderWithStatus,
-} from '~/server/services/tenders';
+import { db } from '~/server/db';
+import { tenders, type TenderInsert } from '~/server/db/schema/tenders.schema';
 import { uploadFileToS3 } from '~/server/s3/upload';
+
+import {
+  withStatus,
+  type TenderWithStatus,
+  type ActionResult,
+  type TenderFormData,
+} from './tenders.utils';
+
+// Re-export types from utils for convenience
+// Note: computeTenderStatus must be imported directly from './tenders.utils'
+export type { TenderWithStatus, ActionResult, TenderFormData };
+
+// ============================================================================
+// Validation
+// ============================================================================
 
 // Validation schema for tender data
 const tenderSchema = z.object({
@@ -42,21 +52,9 @@ const tenderSchema = z.object({
   extendedDate: z.coerce.date().optional().nullable(),
 });
 
-export interface ActionResult {
-  success: boolean;
-  message: string;
-  id?: number;
-}
-
-export interface TenderFormData {
-  title: string;
-  description?: string;
-  pdfLink?: string | null;
-  pdfName?: string | null;
-  startDate: string;
-  endDate: string;
-  extendedDate?: string | null;
-}
+// ============================================================================
+// Helper Functions
+// ============================================================================
 
 /**
  * Check if the current user has permission to manage tenders
@@ -66,6 +64,224 @@ async function checkTenderPermission(): Promise<boolean> {
   const session = await getServerAuthSession();
   return canManageNotifications(session);
 }
+
+// ============================================================================
+// Data Access Functions
+// ============================================================================
+
+/**
+ * Get all tenders with computed status
+ * Returns both live and archived tenders, sorted by effective deadline
+ */
+export async function getAllTenders(): Promise<TenderWithStatus[]> {
+  const results = await db
+    .select()
+    .from(tenders)
+    .orderBy(
+      sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) DESC`
+    );
+
+  return results.map(withStatus);
+}
+
+/**
+ * Get all live tenders (effective deadline >= today)
+ * Ordered by effective deadline ASC (soonest deadline first)
+ */
+export async function getLiveTenders(): Promise<TenderWithStatus[]> {
+  const results = await db
+    .select()
+    .from(tenders)
+    .where(
+      sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) >= CURRENT_DATE`
+    )
+    .orderBy(
+      sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) ASC`
+    );
+
+  return results.map(withStatus);
+}
+
+/**
+ * Get all archived tenders (effective deadline < today)
+ * Ordered by effective deadline DESC (most recently archived first)
+ */
+export async function getArchivedTenders(): Promise<TenderWithStatus[]> {
+  const results = await db
+    .select()
+    .from(tenders)
+    .where(
+      sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) < CURRENT_DATE`
+    )
+    .orderBy(
+      sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) DESC`
+    );
+
+  return results.map(withStatus);
+}
+
+/**
+ * Get a tender by its ID
+ */
+export async function getTenderById(
+  id: number
+): Promise<TenderWithStatus | null> {
+  const result = await db.query.tenders.findFirst({
+    where: (tender, { eq }) => eq(tender.id, id),
+  });
+
+  if (!result) return null;
+  return withStatus(result);
+}
+
+/**
+ * Get the count of tenders (for pagination)
+ * @param archived - If true, count archived tenders; otherwise count live
+ */
+export async function getTenderCount(
+  archived: boolean = false
+): Promise<number> {
+  const condition = archived
+    ? sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) < CURRENT_DATE`
+    : sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) >= CURRENT_DATE`;
+
+  const result = await db
+    .select({ count: sql<number>`count(*)::int` })
+    .from(tenders)
+    .where(condition);
+
+  return result[0]?.count ?? 0;
+}
+
+/**
+ * Get paginated tenders
+ * @param archived - If true, get archived tenders; otherwise get live
+ * @param page - Page number (1-based)
+ * @param limit - Number of items per page
+ */
+export async function getPaginatedTenders(
+  archived: boolean = false,
+  page: number = 1,
+  limit: number = 10
+): Promise<TenderWithStatus[]> {
+  const offset = (page - 1) * limit;
+
+  const condition = archived
+    ? sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) < CURRENT_DATE`
+    : sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) >= CURRENT_DATE`;
+
+  const results = await db
+    .select()
+    .from(tenders)
+    .where(condition)
+    .orderBy(
+      archived
+        ? sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) DESC`
+        : sql`GREATEST("tenders"."end_date", COALESCE("tenders"."extended_date", "tenders"."end_date")) ASC`
+    )
+    .limit(limit)
+    .offset(offset);
+
+  return results.map(withStatus);
+}
+
+/**
+ * Create a new tender
+ * Validates that endDate > startDate
+ */
+async function createTender(data: TenderInsert): Promise<TenderWithStatus> {
+  // Validate dates
+  if (data.endDate <= data.startDate) {
+    throw new Error('End date must be after start date');
+  }
+
+  if (data.extendedDate && data.extendedDate <= data.endDate) {
+    throw new Error('Extended date must be after end date');
+  }
+
+  const [newTender] = await db.insert(tenders).values(data).returning();
+
+  return withStatus(newTender);
+}
+
+/**
+ * Update an existing tender
+ */
+async function updateTender(
+  id: number,
+  data: Partial<TenderInsert>
+): Promise<TenderWithStatus> {
+  const existingTender = await db.query.tenders.findFirst({
+    where: (tender, { eq }) => eq(tender.id, id),
+  });
+
+  if (!existingTender) {
+    throw new Error('Tender not found');
+  }
+
+  // Merge existing data with updates for validation
+  const newStartDate = data.startDate ?? existingTender.startDate;
+  const newEndDate = data.endDate ?? existingTender.endDate;
+  const newExtendedDate = data.extendedDate ?? existingTender.extendedDate;
+
+  // Validate dates
+  if (newEndDate <= newStartDate) {
+    throw new Error('End date must be after start date');
+  }
+
+  if (newExtendedDate && newExtendedDate <= newEndDate) {
+    throw new Error('Extended date must be after end date');
+  }
+
+  const [updatedTender] = await db
+    .update(tenders)
+    .set(data)
+    .where(eq(tenders.id, id))
+    .returning();
+
+  return withStatus(updatedTender);
+}
+
+/**
+ * Update only the extended date of a tender
+ * Special function for extending deadlines
+ */
+async function updateTenderExtendedDate(
+  id: number,
+  newDate: Date
+): Promise<TenderWithStatus> {
+  const existingTender = await db.query.tenders.findFirst({
+    where: (tender, { eq }) => eq(tender.id, id),
+  });
+
+  if (!existingTender) {
+    throw new Error('Tender not found');
+  }
+
+  // Validate extended date is after end date
+  if (newDate <= existingTender.endDate) {
+    throw new Error('Extended date must be after the original end date');
+  }
+
+  const [updatedTender] = await db
+    .update(tenders)
+    .set({ extendedDate: newDate })
+    .where(eq(tenders.id, id))
+    .returning();
+
+  return withStatus(updatedTender);
+}
+
+/**
+ * Delete a tender by ID
+ */
+async function deleteTender(id: number): Promise<void> {
+  await db.delete(tenders).where(eq(tenders.id, id));
+}
+
+// ============================================================================
+// Server Actions (with authentication)
+// ============================================================================
 
 /**
  * Upload a document for a tender (PDF only)
